Use Signals for Deleting User-Related Data

+++ Django-signals_orm-0x04/messaging/views.py
@@ -2,12 +2,13 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth import logout
 from django.shortcuts import redirect
 from django.contrib.auth.models import User
-from django.shortcuts import render
 
 # Create your views here.
 @login_required
-def delte_user(request):
-    user = request.user
-    logout(request)
-    user.delete()
-    return redirect('home')
+def delete_user(request):
+    if request.method != 'POST':
+        user = request.user
+        logout(request)
+        user.delete()
+        return redirect('home')
+    return redirect('profile')
